gfs2: Reduce size of incore inode
Patchwork-id: 131730
O-Subject: [RHEL6.8 fs PATCH v3 4/4] GFS2: Reduce size of incore inode
Bugzilla: 1215794
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch makes no functional changes. Its goal is to reduce the
size of the gfs2 inode in memory by rearranging structures and
changing the size of some variables within the structure.


rhbz#1215794

+++ fs/gfs2/file.c
@@ -735,7 +735,7 @@ static int do_flock(struct file *file, int cmd, struct file_lock *fl)
 	struct gfs2_inode *ip = GFS2_I(file->f_path.dentry->d_inode);
 	struct gfs2_glock *gl;
 	unsigned int state;
-	int flags;
+	u16 flags;
 	int error = 0;
 	int sleeptime;
